+++ backend/Navtrack.Listener/Protocols/Carscop/CarscopProtocol.cs
@@ -8,6 +8,6 @@ namespace Navtrack.Listener.Protocols.Carscop;
 [Service(typeof(IProtocol))]
 public class CarscopProtocol : TkStarProtocol
 {
-    public override int Port => 7016;
+    public override short Port => 7016;
     public override IEnumerable<byte[]> MessageEnd => new List<byte[]> {new byte[] {0x23}, new byte[] {0x5E}};
 }
