+++ backend/Navtrack.Listener/Protocols/Eelink/EelinkProtocol.cs
@@ -6,6 +6,6 @@ namespace Navtrack.Listener.Protocols.Eelink;
 [Service(typeof(IProtocol))]
 public class EelinkProtocol : BaseProtocol
 {
-    public override int Port => 7021;
+    public override short Port => 7021;
     public override byte[] MessageStart => [0x67, 0x67];
 }
