+++ backend/Navtrack.Listener/Protocols/GlobalSat/GlobalSatProtocol.cs
@@ -7,6 +7,6 @@ namespace Navtrack.Listener.Protocols.GlobalSat;
 [Service(typeof(IProtocol))]
 public class GlobalSatProtocol : BaseProtocol
 {
-    public override int Port => 7038;
+    public override short Port => 7038;
     public override IEnumerable<byte[]> MessageEnd => new[] {new byte[] {0x21}};
 }
