+++ backend/Navtrack.Listener/Protocols/Jointech/JointechProtocol.cs
@@ -6,5 +6,5 @@ namespace Navtrack.Listener.Protocols.Jointech;
 [Service(typeof(IProtocol))]
 public class JointechProtocol : BaseProtocol
 {
-    public override int Port => 7040;
+    public override short Port => 7040;
 }
