+++ backend/Navtrack.Listener/Protocols/Megastek/MegastekProtocol.cs
@@ -7,6 +7,6 @@ namespace Navtrack.Listener.Protocols.Megastek;
 [Service(typeof(IProtocol))]
 public class MegastekProtocol : BaseProtocol
 {
-    public override int Port => 7004;
+    public override short Port => 7004;
     public override IEnumerable<byte[]> MessageEnd => new List<byte[]> {new byte[] {0x0D, 0x0A}, new byte[] {0x21}};
 }
